Remove redundant Sentinel target wrapper
