+++ instrumentation/jedis/jedis-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v3_0/JedisConfiguredTargets.java
@@ -78,7 +78,7 @@ public class JedisConfiguredTargets {
         }
       }
     }
-    return JedisServerTarget.ofSentinels(masterName, endpoints);
+    return RedisServerTarget.ofUnorderedEndpointsAndLogicalName(endpoints, masterName);
   }
 
   public static void setClusterTarget(
